fix repo url
i understand this is not a requirement but i am of the opinion that one less everything (query, db check, redirection, display update...) is best when it comes to building a faster, stronger and healthier experience.

Hahha. Feel free to discard - i didn't go on and check all the project for redirects, i.e. 'older links' beeing used but hey.. could be envisaged

Thanks!